requirements.txt
# requirements.txt

google-generativeai
Pillow
PyMuPDF